Adapt OVAL CVE scanner for Hubble
- MOD: Create new OVAL var dict to account for Ubuntu's package name grouping
- MOD: create_vulns function will now consume OVAL var data
- MOD: Move parser options into configurable oval.yaml within Hubble
- MOD: Script will only run unless configured to do so in top.nova
- MOD: Distro facts are now collected via the salt lib module that Hubble uses
- DEL: Remove parse_version, ArgumentParser, distro, subprocess, and re python
       modules from script as they are no longger needed
- DEL: Remove opts.debug as Hubble has its own method to enable debug logging
- MOD: Move opts.output moved into its own write_report_to_file function
- ADD: Add a check for supported distros
- ADD: Create parse_impact_report function to parse report data into Hubble
       friendly format
- MOD: Script now returns data via Hubble's audit function
- MOD: Add ElementTree element check to all root OVAL data
- MOD: Add result = {} arg to build_impact function to resolve cases where
       result might not get populated
- MOD: Package version comparisons refactored into salt lib module
- DEL: Removed various functions responsible for getting system package lists
       and doing version comparisons
- MOD: Various bug fixes
- MOD: Changed docstrings to double quotes
- MOD: Top docstring modified to include configurable options within Hubble
